Extract abstract BaseRestClient, and RestClientInterface
