+++ src/BlocktrailSDK.php
@@ -27,6 +27,7 @@ use Blocktrail\SDK\Bitcoin\BIP32Key;
 use Blocktrail\SDK\Connection\RestClient;
 use Blocktrail\SDK\Exceptions\BlocktrailSDKException;
 use Blocktrail\SDK\Network\BitcoinCash;
+use Blocktrail\SDK\Connection\RestClientInterface;
 use Blocktrail\SDK\V3Crypt\Encryption;
 use Blocktrail\SDK\V3Crypt\EncryptionMnemonic;
 use Blocktrail\SDK\V3Crypt\KeyDerivation;
@@ -36,7 +37,7 @@ use Blocktrail\SDK\V3Crypt\KeyDerivation;
  */
 class BlocktrailSDK implements BlocktrailSDKInterface {
     /**
-     * @var Connection\RestClient
+     * @var Connection\RestClientInterface
      */
     protected $client;
 
@@ -142,10 +143,17 @@ class BlocktrailSDK implements BlocktrailSDKInterface {
     }
 
     /**
-     * @return  RestClient
+     * @return  RestClientInterface
      */
     public function getRestClient() {
         return $this->client;
+    }
+
+    /**
+     * @param RestClientInterface $restClient
+     */
+    public function setRestClient(RestClientInterface $restClient) {
+        $this->client = $restClient;
     }
 
     /**
